+++ BaseLibS/Api/IRegressionMethod.cs
@@ -3,7 +3,7 @@ using BaseLibS.Param;
 
 namespace BaseLibS.Api{
     public interface IRegressionMethod : INamedListItem{
-		RegressionModel Train(BaseVector[] x, float[] y, Parameters param, int nthreads);
+		RegressionModel Train(BaseVector[] x, double[] y, Parameters param, int nthreads);
         Parameters Parameters { get; }
     }
 }
