+++ NumPluginBase/RegressionRank/CorrelationFeatureRanking.cs
@@ -5,23 +5,29 @@ using BaseLibS.Num;
 using BaseLibS.Num.Vector;
 using BaseLibS.Param;
 
-namespace NumPluginBase.RegressionRank{
-	public class CorrelationFeatureRanking : IRegressionFeatureRankingMethod{
-		public int[] Rank(BaseVector[] x, float[] y, Parameters param, IGroupDataProvider data, int nthreads){
+namespace NumPluginBase.RegressionRank {
+	public class CorrelationFeatureRanking : IRegressionFeatureRankingMethod {
+		public int[] Rank(BaseVector[] x, double[] y, Parameters param, IGroupDataProvider data, int nthreads) {
 			int nfeatures = x[0].Length;
 			double[] s = new double[nfeatures];
-			for (int i = 0; i < nfeatures; i++){
-				float[] xx = new float[x.Length];
-				for (int j = 0; j < xx.Length; j++){
-					xx[j] = (float)x[j][i];
+			for (int i = 0; i < nfeatures; i++) {
+				double[] xx = new double[x.Length];
+				for (int j = 0; j < xx.Length; j++) {
+					xx[j] = x[j][i];
 				}
 				s[i] = CalcScore(xx, y);
 			}
 			return ArrayUtils.Order(s);
 		}
 
-		private static double CalcScore(IList<float> xx, IList<float> yy) { return 1 - Math.Abs(ArrayUtils.Correlation(xx, yy)); }
-		public Parameters GetParameters(IGroupDataProvider data) { return new Parameters(); }
+		private static double CalcScore(IList<double> xx, IList<double> yy) {
+			return 1 - Math.Abs(ArrayUtils.Correlation(xx, yy));
+		}
+
+		public Parameters GetParameters(IGroupDataProvider data) {
+			return new Parameters();
+		}
+
 		public string Name => "Abs(Pearson correlation)";
 		public string Description => "";
 		public float DisplayRank => 0;
