+++ NumPluginSvm/SvmRegression.cs
@@ -6,7 +6,7 @@ using NumPluginSvm.Svm;
 
 namespace NumPluginSvm{
 	public class SvmRegression : IRegressionMethod{
-		public RegressionModel Train(BaseVector[] x, float[] y, Parameters param, int nthreads){
+		public RegressionModel Train(BaseVector[] x, double[] y, Parameters param, int nthreads){
 			ParameterWithSubParams<int> kernelParam = param.GetParamWithSubParams<int>("Kernel");
 			SvmParameter sp = new SvmParameter{
 				kernelFunction = KernelFunctions.GetKernelFunction(kernelParam.Value, kernelParam.GetSubParameters()),
